fix(runtime): resolve main merge conflicts for subagent runtime
